Basic topic persistence
Put topics in the `AudienceRepository` for the moment. Not sure whether
that's the best place, or whether it make sense to split them up later.

The slugging functionality is duplicated in topics and audiences at the
moment. I want to keep it separate for the moment in case the needs are
different.